Add files via upload
This is my latest project in prodigy Info Tech. A stopwatch web application developed using HTML, CSS, JavaScript.
This allows users to easily start , stop , and reset  the timer with a simple and intuitive interface.
I focused on making a user friendly , responsive an ensuring accurate timing and a clear display of elapsed time and lap option to 
see the lapped timing.